table: change type reference from `u32` to `u16`

+++ type.h
@@ -3,7 +3,7 @@
 #include "lex.h"
 #include "shared.h"
 
-typedef u32 htype_t;
+typedef u16 htype_t;
 typedef enum htypeinfo_kind_t htypeinfo_kind_t;
 typedef struct htypeinfo_t htypeinfo_t;
